Adding terminal support using ANSI control characters
This now has the basic terminal support necessary for doing lots of cool things
Now the big challenge is going to be updating the various bits to display and
edit screens of text ala ancient forths

On the good news factor though, having lot's of control sequences mapped out will
be very handy, for example ESC c to clear a screen can also trigger a zeroing of
the page of text.

Many of the color controls are going to introduce subtle differences but it is ok.